pyproject.toml
[tool.poetry]
name = "afriaBot"
version = "0.1.0"
description = ""
authors = ["Sadat Issah <51789921+sadatyussuf@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
tweepy = "^4.10.0"
python-dotenv = "^0.20.0"
Flask = "^2.1.3"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
